Add support for Windows
In Windows the RawFd type is replaced with the RawSocket type, updated
connect_socket and accept_socket functions to either take an
IntoRawFd(unix) or IntoRawSocket(windows).

The socket cast into an int (used by the libtls API) is not entirely
safe, since in theiry the windows socket could be a negative number -
but in practice it has always been a positive handle.