TlsStream: take ownership of underlying stream
This breaks compatibility: from_socket() is renamed connect_socket() and
accept() to accept_socket(), this is aligned with the naming in libtls,
it also makes things more clear. The crate version was bumped to 0.2.0.

Its not always intuitive that dropping the holder of a file descriptor
will close the socket, so I used a generic to enable keeping the inner
stream object inside the TlsStream. connect_socket() and accept_socket()
return TlsStream<()>.

connect() and accept() now take ownership of types that implement
AsRawFd/AsRawSocket. This "inner_stream" is stored inside the TlsStream.
The inner() and inner_mut() methods provide access to the object.